refactor(editor): move editor.js to service application
