+++ tsconfig.json
@@ -21,6 +21,7 @@
       "@/presentation/*": ["presentation/*"],
       "@/presentation": ["presentation"],
       "@/domain/*": ["domain/*"],
+      "@/utils": ["utils/*"],
       "@/domain": ["domain"],
       "@/application/*": ["application/*"],
       "@/application": ["application"],
